==> 1_area_code_clean.ipynb <==
import pandas as pd
import zipfile
import pyspark
import pyspark.sql.functions as F
from pyspark.sql import SparkSession
spark = SparkSession.builder.appName('area_code_clean').getOrCreate()

Let's load the dataset.

#path to the zipped dataset
zip_data = 'source_data/full_area_code_dataset.zip'

#extract zipped file
with zipfile.ZipFile(zip_data, 'r') as zip_ref:
    zip_ref.extractall('source_data')

npa_df = spark.read.json("source_data/full_area_code_dataset.json")

The dataset contains NPA numbers (area codes) from Canada as well so lets look at just the United States data.

npa_df.where(npa_df['country'] == "United States").show(10)

Let's remove some of the unecessary columns and pick just the ones we are interested in.

npa_df = npa_df.select('city', 'countryISO', 'npa', 'stateISO', 'zipCode')\
    .where(npa_df['country'] == "United States")

npa_df.show()

Let's see how many rows are in this dataframe.

npa_df.count()

Yikes, that's a lot of data!  I wonder if there is redundant data now that we removed the unnecessary columns.

npa_df.select(F.countDistinct("city")).show()
npa_df.select(F.countDistinct("npa")).show()
npa_df.select(F.countDistinct("countryISO")).show()
npa_df.select(F.countDistinct("zipCode")).show()

Looks like there is a lot of duplicate entries so there is a lot to trim.  Let's run the .distinct() method on the dataframe to remove all the duplicate rows.

npa_df = npa_df.distinct()

Let's see how much smaller the dataframe is now.

npa_df.count()

There is one issue.  Some zip codes have multiple NPA numbers (area codes) associated with them.  

npa_df.orderBy('zipCode', 'npa').show(6)

Lets consolidate those area codes into a list so that we don't have duplicate zip code entries.  We are grouping the zipcodes and cities, aka removing the redundancies and we are performing an aggregate function on the "npa" column.  The function collect_list() aggregates the values into a list and we are renaming this aggregate column as "npa" via the alias() function.

npa_df = npa_df.select("city","zipCode", "npa").groupBy("zipCode", "city")\
    .agg(F.collect_list("npa").alias("npa_list"))

npa_df.show(5)

Much better.  Lets export this data so we can use it in the customer_etl file.

Spark was giving me an error when I tried to save the spark dataframe to a json file so I converted it to a pandas dataframe and saved it using a pandas method.

pandas_df = npa_df.toPandas()
pandas_df.to_json("source_data/area_codes.json", orient='records', lines=True)

# I had to use ', orient='records', lines=True"' because the pandas .to_json function 
# saves the columns as a dictionairy entry in the json file but pyspark read.json function 
# expects each row of the dataframe to be a single line in the json file.
