==> 1_credit_card_etl.ipynb <==
------------------
# Extraction

import pyspark
import pyspark.sql.functions as F
from pyspark.sql import SparkSession
spark = SparkSession.builder.appName('credit_card_etl').getOrCreate()

Let's load cdw_sapp_custmer.json

cc_df = spark.read.json("source_data/cdw_sapp_credit.json")

--------------------------------
# Exploratory Analysis

cc_df.printSchema()
cc_df.show(10)

How many rows do we have in total in this dataframe?

cc_df.count()

cc_df.describe().show()

From these results we can conclude there isn't any null or missing data because each column has the same amount of rows as the entire dataframe and all the min values falls within the logical range.  Also, we can see CC numbers are 16 digits long, SSNs are 9 digits long, Months are 12 or less, and days are 28 or less (guess no one shopped on the 29th, 30th, or 31st).

Let's make sure the CUST_SSN are tied to a single CREDIT_CARD_NO.  when we group by each of these columns they should result in the same number of items.

cc_df.select(F.countDistinct("CREDIT_CARD_NO")).show()
cc_df.select(F.countDistinct("CUST_SSN")).show()
cc_ssn = cc_df.select("CREDIT_CARD_NO","CUST_SSN").groupBy("CREDIT_CARD_NO","CUST_SSN").count()
cc_ssn.count()

Phew!  Things look good there, surprisingly (given how inconsistent the data has been so far...)

Let's see what kind of values are in the TRANSACTION_TYPE column.

cc_df.groupBy('TRANSACTION_TYPE').count().orderBy(F.col('count').desc()).show()

6683 test transactions?? hmm, I wonder what that means.  Surely those can't be trial runs of a transactions..

---------------
# Transforming

Let's move on to the date.  We are required to format it in a YYYYMMDD format.  

In order to do so we need to add leading zeroes to the single digit days and months.  We will use the .lpad function to pad the digit with a specified value, in our case a "0".

cc_df = cc_df.withColumn("MONTH", F.lpad(cc_df.MONTH, 2, '0'))
cc_df = cc_df.withColumn("DAY", F.lpad(cc_df.DAY, 2, '0'))

Let's create the TIMEID column now.

cc_df = cc_df.withColumn('TIMEID', F.format_string("%s%s%s", 
            cc_df['YEAR'], cc_df['MONTH'], cc_df['DAY']))

cc_df.select('DAY','MONTH','YEAR','TIMEID').show(5)

Looks good.  

Let's drop the YEAR, DAY, and MONTH columns since we will not be transfering them into the database.

cc_df = cc_df.drop("DAY")
cc_df = cc_df.drop("MONTH")
cc_df = cc_df.drop("YEAR")

Let's reorder the columns in the order in which they will be loaded into the DB.

cc_df = cc_df.select('TRANSACTION_ID','CREDIT_CARD_NO','TIMEID','CUST_SSN',
                     'BRANCH_CODE','TRANSACTION_TYPE','TRANSACTION_VALUE')

Lets export this data so we can load it into a database.  

Spark was giving me an error when I tried to save the spark dataframe to a json file so I converted it to a pandas dataframe and saved it using a pandas method.

pandas_df = cc_df.toPandas()
pandas_df.to_json('clean_data/credit_card.json')
