==> 1_customer_etl.ipynb <==
------------------
# Extraction

import pyspark
import pyspark.sql.functions as F
from pyspark.sql import SparkSession
spark = SparkSession.builder.appName('customer_etl').getOrCreate()

Let's load cdw_sapp_custmer.json

cust_df = spark.read.json("source_data/cdw_sapp_custmer.json")

--------------------------------
# Exploratory Analysis & Tranforming

cust_df.printSchema()
cust_df.show(10)

How many rows do we have in total in this dataframe?

cust_df.count()

Let's rearrange the columns so they make a bit more sense when looking at them.

cust_df = cust_df.select('CREDIT_CARD_NO', 'SSN', 'CUST_EMAIL', 'CUST_PHONE',\
            'FIRST_NAME','MIDDLE_NAME', 'LAST_NAME', 'APT_NO', 'STREET_NAME',\
            'CUST_CITY', 'CUST_ZIP', 'CUST_STATE', 'CUST_COUNTRY', 'LAST_UPDATED')

cust_df.show(5)

cust_df.describe().show()

From the above summary we can see that the credit card numbers are all 16 digits long, SSN numbers are all 9 digits long, zip codes are all 5 digits long, and all the customers are from the United States.  The only issue is that the phone numbers are only 7 digits long but should be 10 digits. 

It seems like the CUST_ZIP and CUST_STATE columns are in string format so the min and max values might only consider the first character when considering order of values.  Lets confirm that the states are 2 characters in length and the zip codes are all 5 digits long.

cust_df.withColumn("zip_len", F.length(cust_df["CUST_ZIP"]))\
    .groupBy("zip_len").count().show()

cust_df.withColumn("state_len", F.length(cust_df["CUST_STATE"]))\
    .groupBy("state_len").count().show()

Lets make sure the emails are all valid. All have a "@" and a "." (for example, ".com") and no blank spaces.

email_filter_df = cust_df.filter(~F.col('CUST_EMAIL').rlike('^\S+@\S+\.\S+$'))
email_filter_df.select('CUST_EMAIL').show()

#^\S+@\S+\.\S+$ checks to see if the string is a valid email address.  the "^\S+" checks
#if there are one or more non-whitespace characters at the start of the string.  Can't have
#blank spaces in an email address. The ^ is necessary because "ex ample@example.com" would be
#a match on "ample@example.com".  We check for "@" with one or more non-whitespace character after it
#followed by a "." (\. needs to be used because just a "." in regex matches to any single character).
#Finished off by "\S+$" which matches one or more non-whitespace characters that end the string.  
#The "$" is necessary for the same reason that the "^" is necessary.  The "~" is like a NOT so only
#the non matches will be shown.

The customer cities look bunched up, aka two word cities have no spaces between the words.  Let's fix that.

cust_df = cust_df.withColumn('CUST_CITY', 
        F.regexp_replace(cust_df['CUST_CITY'], "(?<=.)([A-Z])", ' $1'))
cust_df.select('CUST_CITY').show(5)

#(?<=.) is a positive lookbehind.  It checks if there is any character behind ([A-Z]).  Don't
#want to match the capital letters in the beginning of the string.
#([A-Z]) looks for any capital letter.  In parantheses because we want to capture this group.
#' $1' Blank space plus the first, and only, captured group. aka this returns the matched 
#capital letter, otherwise the matched capital letter would have been replaced.

Let's make sure all the CC numbers, SSNs, Phone numbers, and Emails are unique. Remember, there are 952 total entries.

cust_df.select(F.countDistinct("CREDIT_CARD_NO")).show()
cust_df.select(F.countDistinct("SSN")).show()
cust_df.select(F.countDistinct("CUST_PHONE")).show()

Hmmm, the Phone numbers don't seem to be unique.  Let's explore further.

cust_df.groupBy('CUST_PHONE').count().orderBy(F.col('count').desc()).show(10)

Let's look at a sample phone number to see if any of the other customer info overlaps.

cust_df.where(cust_df['CUST_PHONE'] == '1236886').show()

In this sample it seems like 3 different people have the same number. I would consider using an area code that corresponds to the customer's location but in this example two of the customers are in the same state and fall under the same area code.

Let's add an area code to the phone numbers based on the customer's zip code and see if that helps differentiate the data.  First, lets load the area code data into a dataframe.  I cleaned up and created this file in the "area_code_data" notebook file.

area_code_df = spark.read.json("source_data/area_codes.json")
area_code_df.show(5)

Since there can be multiple NPA numbers (area codes) for some individual zip codes and there are multiple customers from the same zip code, I decided to alternate the distribution of area codes.  When I would take an area code from the list in the npa_list column, I would move it to the back of the list so that I could just pull the first value in that list for every subsequent query.

To achieve this, I first converted the area_code_df to a pandas dataframe.  I tried to implement this in a spark dataframe at first but it was running way too slow.  The reason being, in order to update the list in the npa_list column of the spark df, I had to update the entire column and dataframe, even though I just needed to update one cell.  Pandas allows for single cell updates and at a much faster speed.  The spark dataframe took 20 minutes and was only 80% completed before it got hung up indefinitely.  The same task in pandas only took 7 seconds to complete.

In the below code I am interating through each row of the cust_df spark dataframe and looking at the zip code in that row.  Then I am looking up the zipcode in the pandas dataframe.  If it finds a match then I am adding the first npa from the npa_list list to a running list called area_code_list.  This list will later get added to the spark dataframe which had to be done this way because you can change a row value in a column one row at a time, in spark dfs you have to update the entire dataframe every time you want to make a change to it.  Then I place the first npa number to the end of the list and save that reordered list to the pandas df using this line: ```npa_list.append(npa_list.pop(0))```

There is also an issue where some users in the customer data have a zip code that doesn't exist in the area_code dataset.  I assigned a ```None``` value to those users for now.

area_code_pd = area_code_df.toPandas()

# This empty list will contain the chosen NPA number for each row.  It will later be joined to 
# the spark dataframe as a new column.
area_code_list = []

# We are taking each row of the customer dataframe and extracting the zipcode and city for the 
# chosen customer.  
for row in cust_df.collect():  #have to use collect.  collect() returns a list of rows.  you can 
    # iterate over this list.  if you don't use the collect() it will return a column object which 
    # can't be iterated over.
    
    zip_code = row[10]
    city = row[9]

    # Here we are saving the npa_list list into a npa_list variable and then taking the first 
    # object in that list and appending it to the running list we created before the for loop.
    # Since there are some customer's with zip codes that don't match the area_code dataframe
    # we try to find a matching NPA number based on their city.
    try:
        npa_list = area_code_pd.loc[area_code_pd['zipCode'] == zip_code, 'npa_list'].iloc[0]
        area_code_list.append(npa_list[0])
        index_pos = area_code_pd[area_code_pd['zipCode'] == zip_code].index[0]
    except:
        npa_list = area_code_pd.loc[area_code_pd['city'] == city, 'npa_list'].iloc[0]
        area_code_list.append(npa_list[0])
        index_pos = area_code_pd[area_code_pd['city'] == city].index[0]

    # Below two lines reorders the python list of npa nums so that the first num goes to the back
    # of the list.  The next line assigns this reordered list to the position of the original
    # npa list in the pandas dataframe.
    npa_list.append(npa_list.pop(0))
    area_code_pd.at[index_pos, 'npa_list'] = npa_list

Let's make sure the list is the same length as the customer dataframe (952) so there aren't any errors when we add it as a column.

len(area_code_list)

Let's join the list of NPA numbers to the customer dataframe.

#had to use pandas because adding a list in pandas is super easy.  in pyspark I would have to 
#create a dataframe for the python list and add an incrementing index column for both dataframes 
#and then join them on the index column and then drop that column.

pandas_cust = cust_df.toPandas()
pandas_cust['npa'] = area_code_list
cust_df = spark.createDataFrame(pandas_cust)
cust_df.show(10)

Let's look at a sample zip code and see if the NPA numbers were assigned in an alternating fashion.

cust_df.select("CUST_ZIP","FIRST_NAME","npa").where(cust_df['CUST_ZIP'] == "01810").show()

Now let's see how many duplicate numbers there are.

cust_df.groupBy('CUST_PHONE','npa').count().orderBy(F.col('count').desc()).show(10)

A big improvement!  Let's investigate further to see if we can fix the last duplicate phone number.  Let's look at the zip codes for the two 1238106 numbers and then lets see if there is an alternative npa available for that zip code.

cust_df.where(cust_df['CUST_PHONE'] == '1238106').show()
area_code_pd.loc[area_code_pd['zipCode'] == "15317", 'npa_list'].iloc[0]

Looks like '724' is available as an alternative.  Let's change one of the numbers to use this NPA number.

cust_df = cust_df.withColumn('npa',
            F.when((F.col("CUST_ZIP") == "15317") & (F.col("CUST_PHONE") == "1238106"), "724" )\
                .otherwise(F.col('npa')))

cust_df.groupBy('CUST_PHONE','npa').count().orderBy(F.col('count').desc()).show(5)

Nice! No more duplicate phone numbers.

Let's change the phone number to a (XXX)XXX-XXXX format.

cust_df = cust_df.withColumn('CUST_PHONE',
            F.format_string("(%s)%s-%s", cust_df['npa'],
            cust_df['CUST_PHONE'][0:3], cust_df['CUST_PHONE'][4:4]))

cust_df.select('CUST_PHONE').show(5)

Let's look at the emails column to see if they are all unique.

cust_df.select(F.countDistinct("CUST_EMAIL")).show()

Strange, why are there so many duplicate emails?

cust_df.groupBy('CUST_EMAIL').count().orderBy(F.col('count').desc()).show(10)

Let's look at a sample email address and explore further.

cust_df.select('SSN','CUST_EMAIL','CUST_PHONE','FIRST_NAME','MIDDLE_NAME','LAST_NAME').where(cust_df['CUST_EMAIL'] == 'BYork@example.com').show()

Conclusion: I really shouldn't alter emails since the email communication will not function if the email address isn't 100% correct.  

But since two seemingly different people seem to be sharing the same email I can only assume the emails are made up and aren't actually used for communication.  

So, as an exercise in data transformation, lets try to alter the emails so we have less unique email addresses.  My solution:  The emails seems to corresspond to the customer's first initial and last names.  But these combinations seems to overlap so lets differentiate the duplicate combinations by adding in a middle initial to the email address.

First, we will use regex to extract the first letter of the customer's middle name and store it in a new column.

cust_df = cust_df.withColumn('Middle_Initial',F.format_string("%s",cust_df['MIDDLE_NAME'][0:1]))

Let's see how this new column looks and if it matches the middle name.

cust_df.select('MIDDLE_NAME','Middle_Initial').show(5)

Let's insert the Middle Initial into the second position of the email address by using the format_string function.  

cust_df = cust_df.withColumn('CUST_EMAIL', F.format_string("%s%s%s", cust_df['CUST_EMAIL'][1:1],
            cust_df['Middle_Initial'], cust_df['CUST_EMAIL'][2:30]))
cust_df = cust_df.drop("Middle_Initial")

cust_df.select('CUST_EMAIL','FIRST_NAME','MIDDLE_NAME','LAST_NAME').show(5)

Let's see how many duplicate emails there are now.  

cust_df.select(F.countDistinct("CUST_EMAIL")).show()
cust_df.groupBy('CUST_EMAIL').count().orderBy(F.col('count').desc()).show(5)

There seems to be one left.  We could eliminate the redudancy further by adding the second letter of the first or middle name into the email address but I will leave it as is for now.

The next requirement is for the First and Last Names to be in Title Case, aka the first letter needs to be capitalized.  Let's use regex to see if that is the case in our data.

f_name_filter = cust_df.filter(~F.col('FIRST_NAME').rlike('^[A-Z][a-z]+$'))
l_name_filter = cust_df.filter(~F.col('LAST_NAME').rlike('^[A-Z][a-z]+$'))

# ^[A-Z] checks to see if there is a single capital letter at the start of the string.
# [a-z]+$ checks to see if there are only lower case letters after the first letter 
# and those letters also go on till the end of the string.  This eliminates the 
# possible matches of McCourt since the C cuts off the lower case letters.

f_name_filter.select('FIRST_NAME').show()
l_name_filter.select('LAST_NAME').show()

We don't have any names that need to converted but if we did here is how we would convert them.

cust_df = cust_df.withColumn('FIRST_NAME', F.initcap(cust_df['FIRST_NAME']))
cust_df = cust_df.withColumn('LAST_NAME', F.initcap(cust_df['LAST_NAME']))

Next, we need to convert the middle names to lower case.  First let's see if we have any names that need to be converted.

m_name_filter = cust_df.filter(~F.col('MIDDLE_NAME').rlike('^[a-z]+$'))
m_name_filter.select('MIDDLE_NAME').show(10)

Looks like we have plenty of middle names that need to be converted.  

cust_df = cust_df.withColumn('MIDDLE_NAME', F.lower(cust_df['MIDDLE_NAME']))

Let's see if there are any middle names left that need to be converted.

m_name_filter = cust_df.filter(~F.col('MIDDLE_NAME').rlike('^[a-z]+$'))
m_name_filter.select('MIDDLE_NAME').show()

Now lets move on to the Street Address.  The requirments ask us to combine the APT_NO and STREET_NAME into a single string separated by a comman with the APT_NO coming after the STREET_NAME.  This would make sense if APT_NO was an apartment number.  But looking at the STREET_NAME column we can see that a leading house number is omitted from the address.  An street name without a house number is not a valid address.  Furthermore not all addresses in the country have apartment numbers yet all the rows in this data have an APT_NO.  My theory is that the APT_NO is actually supposed to be the house number so it should be included in the beginning of the street address string.  

So I will include the APT_NO as the house number but so as to fulfill the requirements I will also include the APT_NO after the street address.

cust_df.select('STREET_NAME','APT_NO').show(10)

cust_df = cust_df.withColumn('FULL_STREET_ADDRESS', F.format_string("%s %s, Apt %s", 
                    cust_df['APT_NO'], cust_df['STREET_NAME'], cust_df['APT_NO']))

cust_df.select('STREET_NAME','APT_NO','FULL_STREET_ADDRESS').show(10, truncate=False)

Let's drop some unecessary rows in preparation for loading into the DB.

cust_df = cust_df.drop("APT_NO")
cust_df = cust_df.drop("STREET_NAME")
cust_df = cust_df.drop("npa")

Let's take one final look at the changes that the dataframe has undergone.

cust_df.show(10)

Let's reorder the columns in the order in which they will be loaded into the DB.

cust_df = cust_df.select('SSN','FIRST_NAME','MIDDLE_NAME','LAST_NAME','CREDIT_CARD_NO',
               'FULL_STREET_ADDRESS','CUST_CITY','CUST_STATE','CUST_COUNTRY','CUST_ZIP',
               'CUST_PHONE','CUST_EMAIL','LAST_UPDATED')

Lets export this data so we can load it into a database.  

Spark was giving me an error when I tried to save the spark dataframe to a json file so I converted it to a pandas dataframe and saved it using a pandas method.

pandas_df = cust_df.toPandas()
pandas_df.to_json('clean_data/customer.json')
