==> 4_Loan_Data_API_ETL.ipynb <==
import requests
from pyspark.sql import SparkSession
import pyspark.sql.functions as F
from pyspark.sql.types import *
import mysql.connector as msql
from mysql.connector import Error
# This file contains our login information for the MySQL server
import credentials as C

# Extraction

Lets extra loan application data from the api endpoint and store it in a Spark dataframe.  In order to store read a json file with spark we must parallelize the json object first.

url = 'https://raw.githubusercontent.com/platformps/LoanDataset/main/loan_data.json'

# assign the headers- not always necessary, but something we have to do with the GitHub API
headers = {'Accept': 'application/vnd.github.v3+json'}

r = requests.get(url, headers=headers)
api_results = r.json()
print(f"Status code: {r.status_code}")

spark = SparkSession.builder.appName("Loan Data Load").getOrCreate()

# Convert the API response to an RDD
json_rdd = spark.sparkContext.parallelize([api_results])
loan_data_df = spark.read.json(json_rdd)
print("Dataframe succefully created")

Lets take a quick look at what the data looks like.

loan_data_df.printSchema()
loan_data_df.show(10)

Lets see what the unique values are in each of the dataframe columns.  This will help determine what schema to use when loading the data into the database.

for column in loan_data_df.columns:
    # skip the first column since that column is all unique values
    if column == 'Application_ID':
        continue
    unique_values = loan_data_df.select(column).distinct().collect()
    print(f"Column: {column}")
    for row in unique_values:
        print(row[column])
    print()

Lets connect to the database to make sure the connection is working.

try:
    conn = msql.connect(host = C.host_name, database = 'creditcard_capstone',
                        user = C.user_name, password = C.password)
    if conn.is_connected():
        cursor = conn.cursor()
        cursor.execute("select database();")
        record = cursor.fetchone()
        print("You're connected to database: ", record)
except Error as e:
    print('Error while connecting to MySQL',e)

Let's construct the sql query to create the table.

loan_table = (
"CREATE TABLE IF NOT EXISTS `cdw_sapp_loan_application` ("
"  `ID` CHAR(8) NOT NULL,"
"  `Gender` VARCHAR(6) NULL,"
"  `Married` VARCHAR(3) NULL,"
"  `Dependents` VARCHAR(2) NULL,"
"  `Education` VARCHAR(12) NULL,"
"  `Self_Employed` VARCHAR(3) NULL," 
"  `Credit_History` VARCHAR(1) NULL,"
"  `Property_Area` VARCHAR(10) NULL,"
"  `Income` VARCHAR(8) NULL,"
"  `Application_Status` VARCHAR(1) NULL,"
"  PRIMARY KEY (`ID`))"
"ENGINE = InnoDB  ")

Time to create the table using the query from above

cursor.execute('DROP TABLE IF EXISTS cdw_sapp_loan_application;')
cursor.execute(loan_table)

# Define a function to insert rows into MySQL
def insert_row(row):
    conn = msql.connect(host = C.host_name, database = 'creditcard_capstone',
                        user = C.user_name, password = C.password)
    cursor = conn.cursor()
    
    insert_query = "INSERT INTO cdw_sapp_loan_application VALUES (%s, %s, %s, %s, %s, %s, %s, %s, %s, %s)"
    values = (row.Application_ID, row.Gender, row.Married,
              row.Dependents, row.Education, row.Self_Employed,
              row.Credit_History, row.Property_Area, row.Income,
              row.Application_Status)
    
    cursor.execute(insert_query, values)
    
    conn.commit()
    cursor.close()
    conn.close()

# Iterate over the PySpark DataFrame and insert rows into MySQL
loan_data_df.foreach(insert_row)
